+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* Add support for Legacy Indie Providers v1
+
 ## 2.0.0
 
 * Rewrite entire package